infra/DA-0000/Fix Container Build
This commit updates the Dockerfile to use the same base image for all
build steps. This is a reminder to also pin your base images.

We were, for some reason, using different images for building ZetaSQL
and running it. Those images have since diverged enough to cause linked
library issues at runtime.

No release will be made for this change. A working build was retagged
with `2020.03.2.3` and `latest`. The original broken container has been
kept as `bad-2020.03.2.3`.

The python library is unaffected.